High Resolution Graphics
Tiles (fonts and graphics) are now regenerated from a high-resolution
image every time the window is resized. This allows us to perfectly
fit the scene into the window, leaving no black padding on the sides.

To access all pixels in a HiDPI setting, we also switched from
rendering surfaces with `SDL_UpdateWindowSurface`, to
rendering textures with `SDL_RenderPresent`.

The window is a mosaic of tiles which are either `N` or `N+1` pixels
wide, and either `M` or `M+1` pixels tall. The large PNG is loaded
once into a surface. When the window size changes, 4 textures are
regenerated with copies of it, downscaled to the four possible
combinations of tile width and height. Each tile gets scaled down
separately.

The downscaling is performed in linear color space, rather than in
gamma-compressed space which causes dimming (average of a black
pixel and a white pixel should be 50% grey = 187, not 128). See
https://blog.johnnovak.net/2016/09/21/what-every-coder-should-know-about-gamma/#image-resizing
for rationale. For performance, we assume a gamma of 2.0.

The tile is logically split into 16 parts (sliced by 3 vertical
lines and 3 horizontal ones), so that:

- the four corners can be positioned to preserve the tile's aspect
  ratio, possibly filling the blank spaces at the top and bottom;
- the top, bottom, left, right, and center areas can be independently
  aligned with output pixels to improve sharpness.

Some tiles, like walls and doors, always stretch to fill the space
regardless of aspect ratio. Other tiles must preserve aspect ratio,
but we allow up to 20% stretch for graphic tiles and 40% for text
tiles.

For text tiles (letters, digits, punctuation signs), the characters'
x-height and baseline (corresponding to the top and bottom of "x")
are pixel-aligned regardless of outlines. This ensures that all
letters neatly line up without jumping up and down. We also reduce
perceived boldness by applying a custom brightness curve; it makes
the text more legible at the smaller sizes.

Optimum pixel alignment is discovered by brute-force, trying all
possible value in 0.1 pixel increment (up to 0.5 pixel). Because this
is a time-consuming process, it is done only once and saved to disk
(`tiles.bin`, committed to git along `tiles.png`). If you ever modify
`tiles.png`, you should regenerate `tiles.bin`: simply delete the
existing file then restart Brogue.

Wall tops are diagonal sine waves, approximately 4 pixels apart.

If the floor tile is completely blank, we fill it with dust-like noise
(isolated dots). This isn't needed for the official Brogue CE tiles,
but gives a great, resolution-dependent solution when using custom
tiles without the Hop-o'-My-Thumb / Pacman-inspired floor markers.

If the window is so large that tiles need to be over 64x64 pixels,
we generate a single, large texture instead of four and use it for
all tiles, allowing the renderer to do some linear interpolation.

To ensure compatibility with older OpenGL drivers, texture dimensions
are always powers of 2.

OpenGL drivers don't like alternating between different textures
too much, so we first draw the background colors then do 4 passes
over the tiles, one per texture.

Some video drivers are quite inefficient, notably in virtual machines,
so there is a new `--no-gpu` command-line parameter to disable
hardware acceleration. The software renderer does not support
HiDPI, though.